PO uint24 -> uint32

+++ contracts/ProfessionalOfficesImpl.sol
@@ -15,17 +15,17 @@ import "./ProfessionalOfficesStorage.sol";
 contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRole { 
 
     // EVENTS
-	event ProfessionalOfficeCreated(uint24 indexed id, bytes32 indexed name);
-	event ProfessionalOfficeActivated(uint24 indexed id, bytes32 indexed name);
-	event ProfessionalOfficeLocked(uint24 indexed id, bytes32 indexed name);
-	event ProfessionalOfficeClosed(uint24 indexed id, bytes32 indexed name);
+	event ProfessionalOfficeCreated(uint32 indexed id, bytes32 indexed name);
+	event ProfessionalOfficeActivated(uint32 indexed id, bytes32 indexed name);
+	event ProfessionalOfficeLocked(uint32 indexed id, bytes32 indexed name);
+	event ProfessionalOfficeClosed(uint32 indexed id, bytes32 indexed name);
 
 	/**
 	 * For manufacturer updates :
 	 *		- manufacturerId is valid and exists
 	 *		- authorized for admin or manufacturer owners if manufacturer status is Activ 
 	 */
-	modifier onlyAuthorizedForUpdate(uint24 _id) {
+	modifier onlyAuthorizedForUpdate(uint32 _id) {
 		require(_id != 0 && _id <= officeSequence, "Invalid id");
 		require(offices[_id].id != 0, "Not found");
 		require(isWhitelistAdmin(msg.sender) || (offices[_id].owners[msg.sender] == true 
@@ -36,7 +36,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
 	/**
 	 * ManufacturerId is valid and exists
 	 */
-	modifier onlyExistingOffice(uint24 _id) {
+	modifier onlyExistingOffice(uint32 _id) {
 		require(_id != 0 && _id <= officeSequence, "Invalid id");
 		require(offices[_id].id != 0, "Not found");
 		_;
@@ -58,7 +58,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
 	/// @return name
 	/// @return status
 	/// @return statusTime 
-	function getProfessionalOffice(uint24 _id) external view onlyExistingOffice(_id) returns(bytes32 name, uint8 status, uint statusTime) {
+	function getProfessionalOffice(uint32 _id) external view onlyExistingOffice(_id) returns(bytes32 name, uint8 status, uint statusTime) {
 		return (offices[_id].name, offices[_id].status, offices[_id].statusTime);
 	}
 
@@ -74,8 +74,8 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
     /// @param _name The office name
     /// @param _owners The owners addresses
     /// @param _technicians The technicians addresses
-    /// @return uint24 id     
-    function addProfessionalOffice(bytes32 _name, address[] calldata _owners, address[] calldata _technicians) external onlyWhitelistAdmin returns(uint24) {
+    /// @return uint32 id     
+    function addProfessionalOffice(bytes32 _name, address[] calldata _owners, address[] calldata _technicians) external onlyWhitelistAdmin returns(uint32) {
         // TODO test name not empty
         for (uint i = 0; i < _owners.length; i++) {
             require(_owners[i] != address(0), "Invalid owner address");
@@ -105,21 +105,21 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
 
     /// @notice Activate professional office
     /// @param _id The professional office id
-	function setProfessionalOfficeActiv(uint24 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
+	function setProfessionalOfficeActiv(uint32 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
 		_changeOfficeStatus(_id, STATUS_ACTIV);
         emit ProfessionalOfficeActivated(_id, offices[_id].name);
 	}
 
     /// @notice Lock professional office
     /// @param _id The professional office id
-	function setProfessionalOfficeLocked(uint24 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
+	function setProfessionalOfficeLocked(uint32 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
 		_changeOfficeStatus(_id, STATUS_LOCKED);
         emit ProfessionalOfficeActivated(_id, offices[_id].name);
 	}
 
     /// @notice Close professional office
     /// @param _id The professional office id
-    function setProfessionalOfficeClosed(uint24 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
+    function setProfessionalOfficeClosed(uint32 _id) external onlyWhitelistAdmin onlyExistingOffice(_id) {
 		_changeOfficeStatus(_id, STATUS_CLOSED);
         emit ProfessionalOfficeActivated(_id, offices[_id].name);
 	}
@@ -127,7 +127,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
     /// @notice Add a new owner address
     /// @param _officeId The professional office id
     /// @param _address The new owner address
-    function addOwner(uint24 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
+    function addOwner(uint32 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
         require(_address != address(0), "Invalid address");
         require(offices[_officeId].owners[_address] != true, "Owner address already registered for this professional office");
         offices[_officeId].owners[_address] = true;
@@ -136,7 +136,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
     /// @notice Disable an existing owner address
     /// @param _officeId The professional office id
     /// @param _address The new owner address
-    function disableOwner(uint24 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
+    function disableOwner(uint32 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
         require(_address != address(0), "Invalid address");
         require(offices[_officeId].owners[_address] == true, "Address not known or already inactiv");
         offices[_officeId].owners[_address] = false;
@@ -145,7 +145,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
     /// @notice Add a new technician address
     /// @param _officeId The professional office id
     /// @param _address The new technician address
-    function addTechnician(uint24 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
+    function addTechnician(uint32 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
         require(_address != address(0), "Invalid address");
         require(officeIdByActivTechnicianAddress[_address] == 0, "Address already activ");
         offices[_officeId].technicians[_address] = true;
@@ -155,7 +155,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
     /// @notice Disable an existing technician address
     /// @param _officeId The professional office id
     /// @param _address The new technician address
-    function disableTechnician(uint24 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
+    function disableTechnician(uint32 _officeId, address _address) external onlyAuthorizedForUpdate(_officeId) {
         require(_address != address(0), "Invalid address");
         require(offices[_officeId].technicians[_address] == true, "Address not known or already inactiv");
         offices[_officeId].technicians[_address] = false;
@@ -166,7 +166,7 @@ contract ProfessionalOfficesImpl is ProfessionalOfficesStorage, WhitelistAdminRo
 	// PRIVATE
 	// -----------------------------------------------------------------------
 	
-	function _changeOfficeStatus(uint24 _id, uint8 _newStatus) internal {
+	function _changeOfficeStatus(uint32 _id, uint8 _newStatus) internal {
 		require(_id != 0 && _id <= officeSequence, "Invalid professional office id");
 		require(offices[_id].id != 0, "Professional office not found");
 		uint8 _oldStatus = offices[_id].status;
